requirements.txt
# All Python dependencies for the AI Avatar Platform

# ── Web Framework ──
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# ── AI / LLM ──
groq==0.4.2
google-cloud-translate==3.13.0

# ── Voice ──
elevenlabs==0.3.0

# ── HTTP ──
requests==2.31.0
httpx==0.25.2

# ── Utils ──
python-dotenv==1.0.0

# ── SadTalker Dependencies ──
# (installed separately via avatar_video/setup.py)
# torch, torchvision, torchaudio
# numpy==1.26.4
# kornia, einops, safetensors
# imageio, librosa, resampy
# basicsr, facexlib, gfpgan
# face_alignment, dlib